Move grant statement to GlobalRuleRegistryService (#10464)
