+++ shardingsphere-governance/shardingsphere-governance-core/src/test/java/org/apache/shardingsphere/governance/core/registry/RegistryCenterTest.java
@@ -91,6 +91,13 @@ public final class RegistryCenterTest {
     }
     
     @Test
+    public void assertInitNodes() {
+        registryCenter.initNodes();
+        verify(registryCenterRepository).persist("/states/datanodes", "");
+        verify(registryCenterRepository).persist("/states/primarynodes", "");
+    }
+    
+    @Test
     public void assertPersistConfigurations() {
         Map<String, DataSourceConfiguration> dataSourceConfigs = createDataSourceConfigurations();
         Collection<RuleConfiguration> schemaRuleConfigs = createRuleConfigurations();
@@ -151,12 +158,5 @@ public final class RegistryCenterTest {
     public void assertRegisterInstanceOnline() {
         registryCenter.registerInstanceOnline();
         verify(registryCenterRepository).persistEphemeral(anyString(), anyString());
-    }
-    
-    @Test
-    public void assertInitNodes() {
-        registryCenter.initNodes();
-        verify(registryCenterRepository).persist("/states/datanodes", "");
-        verify(registryCenterRepository).persist("/states/primarynodes", "");
     }
 }
